# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bright Library API - کتێبخانەی بڕایت
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterBody {
  name: string;
  username: string;
  password: string;
  department: string;
  badgeCode: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  student: "student",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  username: string;
  department: string;
  badgeCode: string;
  role: UserRole;
  isActive: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserActionResponse {
  success: boolean;
}

export interface Book {
  id: number;
  title: string;
  author: string;
  description: string;
  department: string;
  pdfUrl: string;
  coverImage?: string | null;
  createdAt: string;
  downloadCount: number;
}

export interface CreateBookBody {
  title: string;
  author: string;
  description: string;
  department: string;
  pdfUrl: string;
  coverImage?: string | null;
}

export interface UploadCoverResponse {
  url: string;
}

export interface DownloadResponse {
  pdfUrl: string;
  book: Book;
}

export interface Feedback {
  id: number;
  bookId: number;
  bookTitle: string;
  userId: number;
  userName: string;
  content: string;
  createdAt: string;
}

export interface CreateFeedbackBody {
  bookId: number;
  content: string;
}

export interface DownloadEvent {
  id: number;
  bookId: number;
  bookTitle: string;
  userId: number;
  userName: string;
  userDepartment: string;
  userBadgeCode: string;
  downloadedAt: string;
  isRead: boolean;
}

export interface UnreadCountResponse {
  count: number;
}

export interface DepartmentCount {
  department: string;
  count: number;
}

export interface AdminStats {
  totalUsers: number;
  totalBooks: number;
  totalDownloads: number;
  totalFeedback: number;
  booksByDepartment: DepartmentCount[];
  recentDownloads: DownloadEvent[];
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface AdminFeedbackItem {
  id: number;
  bookId: number;
  bookTitle: string;
  bookDepartment: string;
  userId: number;
  userName: string;
  userDepartment: string;
  content: string;
  createdAt: string;
}

export type ListBooksParams = {
  department?: string;
};

export type ListFeedbackParams = {
  bookId?: number;
};

export type UploadCoverBody = {
  cover: Blob;
};