# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bright Library API - کتێبخانەی بڕایت
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/downloads:
    get:
      operationId: listDownloads
      tags: [admin]
      summary: List all download events (admin only)
      responses:
        "200":
          description: List of download events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DownloadEvent"
        "401":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/notifications/unread-count:
    get:
      operationId: getUnreadNotificationsCount
      tags: [admin]
      summary: Get count of unread download notifications (admin only)
      responses:
        "200":
          description: Unread notification count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnreadCountResponse"
        "401":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/notifications/mark-all-read:
    post:
      operationId: markAllNotificationsRead
      tags: [admin]
      summary: Mark all download notifications as read (admin only)
      responses:
        "200":
          description: All notifications marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard statistics
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all registered users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}/deactivate:
    patch:
      operationId: deactivateUser
      tags: [admin]
      summary: Deactivate a student account (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserActionResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}/reactivate:
    patch:
      operationId: reactivateUser
      tags: [admin]
      summary: Reactivate a student account (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User reactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserActionResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}:
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Delete a student account (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserActionResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/feedback:
    get:
      operationId: listAdminFeedback
      tags: [admin]
      summary: List all feedback across all books (admin only)
      responses:
        "200":
          description: List of all feedback with user and book metadata
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminFeedbackItem"
        "401":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: string
        department:
          type: string
        badgeCode:
          type: string
      required:
        - name
        - username
        - password
        - department
        - badgeCode
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        department:
          type: string
        badgeCode:
          type: string
        role:
          type: string
          enum: [student, admin]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - username
        - department
        - badgeCode
        - role
        - isActive
        - createdAt
    UserActionResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    DownloadEvent:
      type: object
      properties:
        id:
          type: integer
        bookId:
          type: integer
        bookTitle:
          type: string
        userId:
          type: integer
        userName:
          type: string
        userDepartment:
          type: string
        userBadgeCode:
          type: string
        downloadedAt:
          type: string
          format: date-time
        isRead:
          type: boolean
      required:
        - id
        - bookId
        - bookTitle
        - userId
        - userName
        - userDepartment
        - userBadgeCode
        - downloadedAt
        - isRead
    UnreadCountResponse:
      type: object
      properties:
        count:
          type: integer
      required:
        - count
    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        totalBooks:
          type: integer
        totalDownloads:
          type: integer
        totalFeedback:
          type: integer
        booksByDepartment:
          type: array
          items:
            $ref: "#/components/schemas/DepartmentCount"
        recentDownloads:
          type: array
          items:
            $ref: "#/components/schemas/DownloadEvent"
      required:
        - totalUsers
        - totalBooks
        - totalDownloads
        - totalFeedback
        - booksByDepartment
        - recentDownloads
    DepartmentCount:
      type: object
      properties:
        department:
          type: string
        count:
          type: integer
      required:
        - department
        - count
...
    AdminFeedbackItem:
      type: object
      properties:
        id:
          type: integer
        bookId:
          type: integer
        bookTitle:
          type: string
        bookDepartment:
          type: string
        userId:
          type: integer
        userName:
          type: string
        userDepartment:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - bookId
        - bookTitle
        - bookDepartment
        - userId
        - userName
        - userDepartment
        - content
        - createdAt