# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bright Library API - کتێبخانەی بڕایت
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new student
 */
export const RegisterBody = zod.object({
  name: zod.string(),
  username: zod.string(),
  password: zod.string(),
  department: zod.string(),
  badgeCode: zod.string(),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    name: zod.string(),
    username: zod.string(),
    department: zod.string(),
    badgeCode: zod.string(),
    role: zod.enum(["student", "admin"]),
    isActive: zod.boolean(),
    createdAt: zod.coerce.date(),
  }),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  username: zod.string(),
  department: zod.string(),
  badgeCode: zod.string(),
  role: zod.enum(["student", "admin"]),
  isActive: zod.boolean(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List books, optionally filtered by department
 */
export const ListBooksQueryParams = zod.object({
  department: zod.coerce.string().optional(),
});

export const ListBooksResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  author: zod.string(),
  description: zod.string(),
  department: zod.string(),
  pdfUrl: zod.string(),
  coverImage: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  downloadCount: zod.number(),
});
export const ListBooksResponse = zod.array(ListBooksResponseItem);

/**
 * @summary Add a new book (admin only)
 */
export const CreateBookBody = zod.object({
  title: zod.string(),
  author: zod.string(),
  description: zod.string(),
  department: zod.string(),
  pdfUrl: zod.string(),
  coverImage: zod.string().nullish(),
});

/**
 * @summary Get a single book by ID
 */
export const GetBookParams = zod.object({
  id: zod.coerce.number(),
});

export const GetBookResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  author: zod.string(),
  description: zod.string(),
  department: zod.string(),
  pdfUrl: zod.string(),
  coverImage: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  downloadCount: zod.number(),
});

/**
 * @summary Delete a book (admin only)
 */
export const DeleteBookParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteBookResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Record a book download and return the PDF URL
 */
export const DownloadBookParams = zod.object({
  id: zod.coerce.number(),
});

export const DownloadBookResponse = zod.object({
  pdfUrl: zod.string(),
  book: zod.object({
    id: zod.number(),
    title: zod.string(),
    author: zod.string(),
    description: zod.string(),
    department: zod.string(),
    pdfUrl: zod.string(),
    coverImage: zod.string().nullish(),
    createdAt: zod.coerce.date(),
    downloadCount: zod.number(),
  }),
});

/**
 * @summary List all feedback (with optional book filter)
 */
export const ListFeedbackQueryParams = zod.object({
  bookId: zod.coerce.number().optional(),
});

export const ListFeedbackResponseItem = zod.object({
  id: zod.number(),
  bookId: zod.number(),
  bookTitle: zod.string(),
  userId: zod.number(),
  userName: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListFeedbackResponse = zod.array(ListFeedbackResponseItem);

/**
 * @summary Submit feedback for a book
 */
export const CreateFeedbackBody = zod.object({
  bookId: zod.number(),
  content: zod.string(),
});

/**
 * @summary List all download events (admin only)
 */
export const ListDownloadsResponseItem = zod.object({
  id: zod.number(),
  bookId: zod.number(),
  bookTitle: zod.string(),
  userId: zod.number(),
  userName: zod.string(),
  userDepartment: zod.string(),
  userBadgeCode: zod.string(),
  downloadedAt: zod.coerce.date(),
  isRead: zod.boolean(),
});
export const ListDownloadsResponse = zod.array(ListDownloadsResponseItem);

/**
 * @summary Get count of unread download notifications (admin only)
 */
export const GetUnreadNotificationsCountResponse = zod.object({
  count: zod.number(),
});

/**
 * @summary Mark all download notifications as read (admin only)
 */
export const MarkAllNotificationsReadResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get admin dashboard statistics
 */
export const GetAdminStatsResponse = zod.object({
  totalUsers: zod.number(),
  totalBooks: zod.number(),
  totalDownloads: zod.number(),
  totalFeedback: zod.number(),
  booksByDepartment: zod.array(
    zod.object({
      department: zod.string(),
      count: zod.number(),
    }),
  ),
  recentDownloads: zod.array(
    zod.object({
      id: zod.number(),
      bookId: zod.number(),
      bookTitle: zod.string(),
      userId: zod.number(),
      userName: zod.string(),
      userDepartment: zod.string(),
      userBadgeCode: zod.string(),
      downloadedAt: zod.coerce.date(),
      isRead: zod.boolean(),
    }),
  ),
});

/**
 * @summary List all registered users (admin only)
 */
export const ListUsersResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  username: zod.string(),
  department: zod.string(),
  badgeCode: zod.string(),
  role: zod.enum(["student", "admin"]),
  isActive: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const ListUsersResponse = zod.array(ListUsersResponseItem);

/**
 * @summary Deactivate a student account (admin only)
 */
export const DeactivateUserParams = zod.object({
  id: zod.coerce.number(),
});

export const DeactivateUserResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Reactivate a student account (admin only)
 */
export const ReactivateUserParams = zod.object({
  id: zod.coerce.number(),
});

export const ReactivateUserResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Delete a student account (admin only)
 */
export const DeleteUserParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteUserResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Upload a book cover image (admin only)
 */
export const UploadCoverBody = zod.object({
  cover: zod.instanceof(File),
});

/**
 * @summary Request a presigned URL for file upload
 */

export const RequestUploadUrlBody = zod.object({
  name: zod.string().min(1),
  size: zod.number().min(1),
  contentType: zod.string().min(1),
});

export const RequestUploadUrlResponse = zod.object({
  uploadURL: zod.string().url(),
  objectPath: zod.string(),
});

/**
 * @summary Serve an object entity from PRIVATE_OBJECT_DIR
 */
export const GetStorageObjectParams = zod.object({
  objectPath: zod.coerce.string(),
});

/**
 * @summary List all feedback across all books (admin only)
 */
export const ListAdminFeedbackResponseItem = zod.object({
  id: zod.number(),
  bookId: zod.number(),
  bookTitle: zod.string(),
  bookDepartment: zod.string(),
  userId: zod.number(),
  userName: zod.string(),
  userDepartment: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListAdminFeedbackResponse = zod.array(
  ListAdminFeedbackResponseItem,
);